nodes refactor
