+++ blender/nodes/source_object.py
@@ -1,0 +1,19 @@
+import bpy
+
+from . import base
+from ..categories import SOURCE_DATA
+
+
+class ElementsSourceObjectNode(base.BaseNode):
+    bl_idname = 'elements_source_object_node'
+    bl_label = 'Source Object'
+
+    name: bpy.props.StringProperty()
+    category = SOURCE_DATA
+
+    def init(self, context):
+        out = self.outputs.new('elements_struct_socket', 'Object')
+        out.text = 'Source Geometry'
+
+    def draw_buttons(self, context, layout):
+        layout.prop_search(self, 'name', bpy.data, 'objects', text='')
